@@ -18,28 +18,30 @@
     "e2e": "./cypress/automation/run.sh"
   },
   "dependencies": {
-    "@apollo/client": "^3.6.9",
-    "@emotion/react": "^11.10.0",
-    "@emotion/styled": "^11.10.0",
+    "@apollo/client": "^3.10.8",
+    "@apollo/experimental-nextjs-app-support": "^0.11.2",
+    "@emotion/react": "11.11.4",
+    "@emotion/styled": "11.11.5",
     "@faker-js/faker": "^7.6.0",
     "@fontsource/roboto": "^4.5.8",
     "@hookform/resolvers": "^2.9.11",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/lab": "5.0.0-alpha.102",
-    "@mui/material": "^5.10.1",
+    "@mui/icons-material": "5.15.20",
+    "@mui/lab": "5.0.0-alpha.170",
+    "@mui/material": "5.15.20",
+    "@mui/material-nextjs": "5.15.11",
     "@react-oauth/google": "^0.2.8",
     "dnd-core": "^16.0.1",
     "formik": "^2.2.9",
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
     "nanoid": "^4.0.1",
-    "next": "13.1.3",
+    "next": "14.2.4",
     "notistack": "^2.0.8",
     "pluralize": "^8.0.0",
-    "react": "18.2.0",
+    "react": "18.3.1",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "18.2.0",
+    "react-dom": "18.3.1",
     "react-hook-form": "^7.43.5",
     "yup": "^0.32.11",
     "zod": "^3.21.4"
@@ -57,13 +59,13 @@
     "@types/jest": "^29.2.5",
     "@types/node": "18.7.7",
     "@types/pluralize": "^0.0.29",
-    "@types/react": "18.0.17",
-    "@types/react-dom": "18.0.6",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
     "@typescript-eslint/eslint-plugin": "^5.37.0",
     "@typescript-eslint/parser": "^5.37.0",
     "cypress": "^12.5.1",
     "eslint": "8.22.0",
-    "eslint-config-next": "13.1.3",
+    "eslint-config-next": "14.2.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-unused-imports": "^2.0.0",
     "jest": "^29.3.1",
