@@ -1,15 +1,26 @@
 import { GoogleOAuthProvider } from "@react-oauth/google";
-import { useRouter } from "next/router";
-import { PropsWithChildren, useCallback, useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
+import {
+  PropsWithChildren,
+  useCallback,
+  useEffect,
+  useState,
+  useTransition,
+} from "react";
 
 import client from "./client";
 import AuthContext from "./context";
-import storage from "modules/storage";
+import {
+  clearTokens,
+  loadTokens,
+  saveTokens as saveTokensToCookie,
+} from "./tokens";
 
 const AuthProvider = ({ children }: PropsWithChildren) => {
   const [accessToken, setAccessToken] = useState<string | null>(null);
   const [refreshToken, setRefreshToken] = useState<string | null>(null);
   const [authenticated, setAuthenticated] = useState(false);
+  const [, startTransition] = useTransition();
 
   const router = useRouter();
 
@@ -17,9 +28,11 @@ const AuthProvider = ({ children }: PropsWithChildren) => {
     (access_token: string, refresh_token: string) => {
       setAccessToken(access_token);
       setRefreshToken(refresh_token);
-      storage.saveTokens({
-        accessToken: access_token,
-        refreshToken: refresh_token,
+      startTransition(async () => {
+        await saveTokensToCookie({
+          accessToken: access_token,
+          refreshToken: refresh_token,
+        });
       });
       setAuthenticated(true);
     },
@@ -34,6 +47,7 @@ const AuthProvider = ({ children }: PropsWithChildren) => {
       );
       saveTokens(access_token, refresh_token);
       router.push("/");
+      router.refresh();
     },
     [saveTokens, router]
   );
@@ -46,6 +60,7 @@ const AuthProvider = ({ children }: PropsWithChildren) => {
       );
       saveTokens(access_token, refresh_token);
       router.push("/");
+      router.refresh();
     },
     [router, saveTokens]
   );
@@ -55,27 +70,33 @@ const AuthProvider = ({ children }: PropsWithChildren) => {
       const { access_token, refresh_token } = await client.googleSignin(token);
       saveTokens(access_token, refresh_token);
       router.push("/");
+      router.refresh();
     },
     [router, saveTokens]
   );
 
-  const signout = useCallback(() => {
-    setAccessToken(null);
-    setRefreshToken(null);
-    storage.clean();
-    setAuthenticated(false);
+  const signout = () => {
+    startTransition(async () => {
+      await clearTokens();
+      setAccessToken(null);
+      setRefreshToken(null);
+      setAuthenticated(false);
+    });
     router.push("/");
-  }, [router]);
+    router.refresh();
+  };
 
   const googleClientID = process.env.NEXT_PUBLIC_OAUTH_GOOGLE_CLIENT_ID || "";
 
   useEffect(() => {
-    const tokens = storage.loadTokens();
-    if (tokens) {
-      setAccessToken(tokens.accessToken);
-      setRefreshToken(tokens.refreshToken);
-      setAuthenticated(true);
-    }
+    startTransition(async () => {
+      const res = await loadTokens();
+      if (res) {
+        setAccessToken(res.accessToken);
+        setRefreshToken(res.refreshToken);
+        setAuthenticated(true);
+      }
+    });
   }, []);
 
   return (
